# Add comprehensive performance optimization for large datasets and Japanese README