… Gemfile.lock
… QUERY_OPTIMIZATION_IMMEDIATE.md
… README.md
… README_ja.md
… config/database.yml