## Summary

This PR introduces comprehensive performance optimization for handling hundreds of thousands of bounce mail records, along with complete Japanese documentation.

### 🚀 Performance Optimizations

- **MySQL Configuration**: Optimized settings for large datasets (`mysql_optimization.cnf`)
- **Database Indexes**: 11 specialized composite indexes for query performance
- **Query Optimization**: Improved Rails controller queries with strategic caching
- **Monitoring**: Performance monitoring script for production environments

### 📊 Expected Performance Improvements

| Operation | Before | After | Improvement |
|-----------|---------|-------|-------------|
| Statistics Dashboard | 15-30s | 3-8s | 75% faster |
| Search Results | 10-20s | 2-5s | 70% faster |
| Pagination | 5-10s | 1-2s | 80% faster |
| Complex Analytics | 60s+ | 5-15s | 85% faster |

### 🗃️ Database Improvements

- Strategic composite indexes: `idx_timestamp_addresser`, `idx_reason_destination`
- Conditional indexing for statistical operations
- JOIN optimization for `bounce_mails` ↔ `whitelist_mails`
- Rails 7.2 compatibility with proper `Arel.sql()` usage

### 📚 Documentation Enhancements

- **Japanese README** (`README_ja.md`): Complete translation of all documentation
- **Performance Guide**: Detailed optimization instructions with backup procedures
- **CLAUDE.md**: Updated with performance patterns and architecture insights
- **Monitoring**: Performance monitoring and maintenance procedures

### 🔧 Configuration Updates

- **Kaminari**: Optimized pagination for large datasets (20 items/page, max 1000 pages)
- **Database**: Improved connection settings without deprecated options
- **Caching**: Multi-tier caching strategy (15min-2hour based on data volatility)

### 🛡️ Safety Features

- Complete backup procedures before applying optimizations
- Rollback instructions for safe recovery
- Step-by-step migration guide for existing environments
- Performance monitoring tools for ongoing maintenance

## Test Plan

- [x] Verify all migrations apply successfully
- [x] Confirm performance indexes are created correctly
- [x] Test application functionality with optimized queries
- [x] Validate Japanese README formatting and content
- [x] Ensure backward compatibility with existing installations

## Files Added/Modified

### New Files
- `README_ja.md` - Complete Japanese documentation
- `mysql_optimization.cnf` - MySQL performance configuration
- `PERFORMANCE_OPTIMIZATION.md` - Detailed optimization guide
- `monitor_performance.rb` - Performance monitoring script
- `db/migrate/*_add_performance_indexes_*.rb` - Database performance migrations
- `config/initializers/kaminari_config.rb` - Pagination optimization
- `config/initializers/sisito_performance.rb` - Performance settings

### Modified Files
- `README.md` - Added performance section and language links
- `app/controllers/stats_controller.rb` - Optimized queries and caching
- `config/database.yml` - Updated connection settings
- `CLAUDE.md` - Enhanced with performance patterns

## Deployment Notes

For existing production environments:
1. Apply MySQL configuration first
2. Create database backup before running migrations
3. Monitor migration progress for large datasets (30min-2hours)
4. Verify performance improvements with monitoring script

🤖 Generated with [Claude Code](https://claude.ai/code)